Updated based on the Reviews
There are cases that Service bugs due to the request rate of the tests, and do not store the rating objects. In that case there are assumptions to ignore these tests.

+++ src/test/java/com/uwetrottmann/tmdb2/TmdbTestSuite.java
@@ -2,6 +2,8 @@ package com.uwetrottmann.tmdb2;
 
 import com.googlecode.junittoolbox.SuiteClasses;
 import com.googlecode.junittoolbox.WildcardPatternSuite;
+import com.uwetrottmann.tmdb2.annotations.RequiresAccountSession;
+import com.uwetrottmann.tmdb2.annotations.RequiresGuestSession;
 import com.uwetrottmann.tmdb2.entities.Account;
 import com.uwetrottmann.tmdb2.entities.BaseMovie;
 import com.uwetrottmann.tmdb2.entities.FavoriteMedia;
@@ -16,13 +18,18 @@ import org.assertj.core.util.Strings;
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
 import org.junit.ClassRule;
+import org.junit.Test;
 import org.junit.rules.ExternalResource;
 import org.junit.runner.RunWith;
 
 import java.io.IOException;
+import java.lang.reflect.Field;
+import java.lang.reflect.Method;
 import java.text.ParseException;
 import java.util.ArrayList;
+import java.util.Vector;
 
+import static com.uwetrottmann.tmdb2.BaseTestCase.testClassToRun;
 import static com.uwetrottmann.tmdb2.TestData.initializeTestingIntegrityData;
 import static com.uwetrottmann.tmdb2.TestData.testList;
 import static com.uwetrottmann.tmdb2.TestData.testListMovies;
@@ -32,20 +39,26 @@ import static com.uwetrottmann.tmdb2.TestData.testTvSeason;
 import static com.uwetrottmann.tmdb2.TestData.testTvShow;
 
 @RunWith(WildcardPatternSuite.class)
-@SuiteClasses({"**/*Test.class"})
+@SuiteClasses({"**/*.class", "!TmdbTestSuite.class", "!BaseTestCase.class"})
 public class TmdbTestSuite {
 
     private static String API_KEY;
-    public static String USERNAME ;
-    public static String PASSWORD;
+    private static String USERNAME;
+    private static String PASSWORD;
     static final boolean DEBUG = true;
 
     static Tmdb unauthenticatedInstance;
     static Tmdb authenticatedInstance;
 
-    static Boolean initialized = false;
+
+    public static Boolean initialized = false;
+    public static Boolean accountDataInitialized = false;
+    public static Boolean guestDataInitialized = false;
 
     public static Boolean suiteRunning = false;
+
+    private static Account account;
+
     @ClassRule
     public static ExternalResource triggerSuiteRunning = new ExternalResource() {
         @Override
@@ -55,11 +68,10 @@ public class TmdbTestSuite {
     };
 
     @BeforeClass
-    public static void initializeEnvironment() throws IOException, ParseException {
-        if (initialized)
+    public static void initializeEnvironment() throws IOException, ParseException, InterruptedException {
+        if (initialized) {
             return;
-
-        initialized = true;
+        }
 
         if (Strings.isNullOrEmpty(USERNAME) || Strings.isNullOrEmpty(PASSWORD)) {
             USERNAME = System.getenv("TMDB_USERNAME");
@@ -76,9 +88,55 @@ public class TmdbTestSuite {
         unauthenticatedInstance = new BaseTestCase.TestTmdb(API_KEY);
         authenticatedInstance = new BaseTestCase.TestTmdb(API_KEY);
 
+        initializeData();
+
+        initialized = true;
+    }
+
+    public static void initializeData() throws ParseException, IOException, InterruptedException {
         initializeTestingIntegrityData();
-        initializeAccountData();
-        initializeGuestData();
+
+        try {
+            Field field = ClassLoader.class.getDeclaredField("classes");
+            field.setAccessible(true);
+
+            @SuppressWarnings({"unchecked", "rawtypes"})
+            Vector<Class> classes = (Vector<Class>) field.get(BaseTestCase.class.getClassLoader());
+            Boolean classCounted;
+            for (Class<?> clazz : new Vector<>(classes)) {
+
+                if (clazz.getName().startsWith("com.uwetrottmann.tmdb2") && !clazz.getSimpleName().equals("BaseTestCase") && !clazz.getSimpleName().equals("TmdbTestSuite")) {
+                    classCounted = false;
+
+                    if (clazz.isAnnotationPresent(RequiresAccountSession.class) && !accountDataInitialized) {
+                        initializeAccountData();
+                    }
+                    if (clazz.isAnnotationPresent(RequiresGuestSession.class) && !guestDataInitialized) {
+                        initializeGuestData();
+                    }
+
+                    for (Method method : clazz.getDeclaredMethods()) {
+                        if (method.isAnnotationPresent(Test.class)) {
+                            if (!classCounted) {
+                                classCounted = true;
+                                testClassToRun++;
+                            }
+
+                            if (method.isAnnotationPresent(RequiresAccountSession.class) && !accountDataInitialized) {
+                                initializeAccountData();
+                            } else if (method.isAnnotationPresent(RequiresGuestSession.class) && !guestDataInitialized) {
+                                initializeGuestData();
+                            }
+                        }
+                    }
+
+                }
+            }
+        } catch (IllegalAccessException e) {
+            e.printStackTrace();
+        } catch (NoSuchFieldException e) {
+            e.printStackTrace();
+        }
     }
 
     public static void initializeGuestData() throws IOException {
@@ -92,17 +150,20 @@ public class TmdbTestSuite {
         authenticatedInstance.tvService().addRating(testTvShow.id, AuthenticationType.GUEST, ratingObject).execute();
 
         authenticatedInstance.tvEpisodesService().addRating(testTvEpisode.id, testTvSeason.season_number, testTvEpisode.episode_number, AuthenticationType.GUEST, ratingObject).execute();
+
+        guestDataInitialized = true;
     }
 
-    public static void initializeAccountData() throws IOException {
-        if (USERNAME == null || PASSWORD == null)
+    public static void initializeAccountData() throws IOException, InterruptedException {
+        if (Strings.isNullOrEmpty(USERNAME) || Strings.isNullOrEmpty(PASSWORD)) {
             return;
+        }
 
         System.out.println("Found TMDB Credentials. Authenticating Instance...");
 
         authenticatedInstance.accountSession(USERNAME, PASSWORD);
 
-        Account account = authenticatedInstance.accountService().summary().execute().body();
+        account = authenticatedInstance.accountService().summary().execute().body();
 
         ListCreateRequest request = new ListCreateRequest();
         request.name = "TestList";
@@ -135,17 +196,19 @@ public class TmdbTestSuite {
             watchlistMedia.media_id = i;
 
             authenticatedInstance.listsService().addMovie(testList.id, listOperation).execute();
-            authenticatedInstance.accountService().favorite(0, favoriteMedia).execute();
-            authenticatedInstance.accountService().watchlist(0, watchlistMedia).execute();
+            Thread.sleep(200);
+            authenticatedInstance.accountService().favorite(account.id, favoriteMedia).execute();
+            Thread.sleep(200);
+            authenticatedInstance.accountService().watchlist(account.id, watchlistMedia).execute();
         }
 
         favoriteMedia.media_type = MediaType.TV;
         favoriteMedia.media_id = testTvShow.id;
-        authenticatedInstance.accountService().favorite(0, favoriteMedia).execute();
+        authenticatedInstance.accountService().favorite(account.id, favoriteMedia).execute();
 
         watchlistMedia.media_type = MediaType.TV;
         watchlistMedia.media_id = 1399;
-        authenticatedInstance.accountService().watchlist(0, watchlistMedia).execute();
+        authenticatedInstance.accountService().watchlist(account.id, watchlistMedia).execute();
 
         RatingObject ratingObject = new RatingObject();
         ratingObject.value = 7.5;
@@ -156,10 +219,14 @@ public class TmdbTestSuite {
 
         authenticatedInstance.tvEpisodesService().addRating(testTvShow.id, testTvSeason.season_number, testTvEpisode.episode_number, AuthenticationType.ACCOUNT, ratingObject).execute();
 
+
+        Thread.sleep(5000);
+
+        accountDataInitialized = true;
     }
 
     @AfterClass
-    public static void destructEnvironment() throws IOException {
+    public static void destructEnvironment() throws IOException, InterruptedException {
 
         System.out.println("Test Classes to run have been finished. Cleaning up...");
         rollbackAccountData();
@@ -168,6 +235,10 @@ public class TmdbTestSuite {
 
 
     public static void rollbackGuestData() throws IOException {
+        if (!guestDataInitialized) {
+            return;
+        }
+
         authenticatedInstance.moviesService().deleteRating(testMovie.id, AuthenticationType.GUEST).execute();
 
         authenticatedInstance.tvService().deleteRating(testTvShow.id, AuthenticationType.GUEST).execute();
@@ -176,9 +247,10 @@ public class TmdbTestSuite {
     }
 
 
-    public static void rollbackAccountData() throws IOException {
-        if (USERNAME == null || PASSWORD == null)
+    public static void rollbackAccountData() throws IOException, InterruptedException {
+        if (!accountDataInitialized) {
             return;
+        }
 
         FavoriteMedia favoriteMedia = new FavoriteMedia();
         favoriteMedia.media_type = MediaType.MOVIE;
@@ -191,17 +263,18 @@ public class TmdbTestSuite {
             favoriteMedia.media_id = i;
             watchlistMedia.media_id = i;
 
-            authenticatedInstance.accountService().favorite(0, favoriteMedia).execute();
-            authenticatedInstance.accountService().watchlist(0, watchlistMedia).execute();
+            authenticatedInstance.accountService().favorite(account.id, favoriteMedia).execute();
+            Thread.sleep(200);
+            authenticatedInstance.accountService().watchlist(account.id, watchlistMedia).execute();
         }
 
         favoriteMedia.media_type = MediaType.TV;
         favoriteMedia.media_id = testTvShow.id;
-        authenticatedInstance.accountService().favorite(0, favoriteMedia).execute();
+        authenticatedInstance.accountService().favorite(account.id, favoriteMedia).execute();
 
         watchlistMedia.media_type = MediaType.TV;
         watchlistMedia.media_id = 1399;
-        authenticatedInstance.accountService().watchlist(0, watchlistMedia).execute();
+        authenticatedInstance.accountService().watchlist(account.id, watchlistMedia).execute();
 
         authenticatedInstance.moviesService().deleteRating(testMovie.id, AuthenticationType.ACCOUNT).execute();
